Ea_OS/Organs/BIOwerk-main/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "biowerk-matrix"
version = "0.1.0"
description = "Shared library for BIOwerk microservices - Core utilities, models, auth, database, and LLM integration"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "BIOwerk Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.6",
    "pydantic>=2.9.2",
    "blake3>=0.4.1",
    "httpx>=0.27.2",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.13.3",
    "motor>=3.6.0",
    "redis>=5.2.1",
    "pydantic-settings>=2.6.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "bcrypt>=4.2.1",
    "cryptography>=44.0.0",
    "safety>=3.2.11",
    "openai>=1.58.1",
    "anthropic>=0.39.0",
    "ollama>=0.4.4",
    "llama-cpp-python>=0.3.2",
    "huggingface-hub>=0.27.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-instrumentation-httpx>=0.48b0",
    "opentelemetry-instrumentation-asyncpg>=0.48b0",
    "opentelemetry-instrumentation-redis>=0.48b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "opentelemetry-exporter-jaeger>=1.27.0",
    "opentelemetry-exporter-prometheus>=0.48b0",
    "opentelemetry-semantic-conventions>=0.48b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.23.8",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.6.1",
    "pytest-timeout>=2.3.1",
    "black>=24.10.0",
    "isort>=5.13.2",
    "flake8>=7.1.1",
    "pylint>=3.0.0",
    "mypy>=1.11.0",
    "bandit>=1.7.10",
    "pre-commit>=3.8.0",
    "types-redis>=4.6.0",
    "types-passlib>=1.7.7",
]

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | alembic/versions
)/
'''

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/alembic/versions/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=matrix",
    "--cov=services",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "security: marks tests as security tests",
]

[tool.coverage.run]
source = ["matrix", "services"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/alembic/versions/*",
    "*/venv/*",
    "*/.venv/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
fail_under = 80
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@abc.abstractmethod",
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true

[[tool.mypy.overrides]]
module = [
    "motor.*",
    "blake3.*",
    "ollama.*",
    "anthropic.*",
    "openai.*",
    "llama_cpp.*",
]
ignore_missing_imports = true

[tool.pylint.main]
py-version = "3.10"
jobs = 0
ignore-paths = [
    "^.*/alembic/versions/.*$",
    "^.*/tests/.*$",
]

[tool.pylint.messages_control]
disable = [
    "C0103",  # Invalid name
    "C0114",  # Missing module docstring
    "C0115",  # Missing class docstring
    "C0116",  # Missing function docstring
    "R0903",  # Too few public methods
    "R0913",  # Too many arguments
    "W0212",  # Access to protected member
    "E0401",  # Import error (handled by mypy)
]

[tool.pylint.format]
max-line-length = 120

[tool.pylint.design]
max-args = 10
max-attributes = 15
max-bool-expr = 5
max-branches = 15
max-locals = 20
max-parents = 7
max-public-methods = 25
max-returns = 6
max-statements = 60
min-public-methods = 1

[tool.bandit]
exclude_dirs = [
    "tests",
    "test_*.py",
    ".venv",
    "venv",
]
skips = ["B101"]  # Skip assert_used check
